pyproject.toml
[tool.poetry]
name = "sourcery-rules-generator"
version = "0.1.0"
description = ""
authors = ["reka <reka@hey.com>"]

[tool.poetry.scripts]
sourcery-rules = "sourcery_rules_generator.cli:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "0.7.0"}
rich = "^12.6.0"
pydantic = "^1.10.2"
ruamel-yaml = "^0.17.21"

[tool.poetry.dev-dependencies]
pytest = "7.1.3"
pytest-mock = "3.8.2"
coverage = "6.4.4"
black = "22.8.0"
isort = "5.10.1"
mypy = "0.971"
flake8 = "5.0.4"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
